mcp-server-app/mcp-server/requirements.txt
mcp>=1.2.0
httpx>=0.28.0
python-dotenv>=1.0.0
